backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
django-environ==0.11.2
black==23.12.1
mypy==1.7.1
django-stubs==4.2.7
django-extensions==3.2.3
setuptools>=65.5.0
drf-yasg==1.21.7
factory-boy==3.3.0

